SecureComm/services/backend/pyproject.toml
[project]
name = "securecomm-backend"
version = "0.1.0"
description = "SecureComm FastAPI backend for store-and-forward messaging."
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.5",
    "gunicorn==22.0.0",
    "sqlalchemy==2.0.32",
    "alembic==1.13.2",
    "pyjwt[crypto]==2.9.0",
    "psycopg[binary,pool]==3.2.1",
    "redis==5.0.8",
    "python-dotenv==1.0.1"
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.2",
    "httpx==0.27.2",
    "ruff==0.5.7",
    "pytest-asyncio==0.23.8"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["securecomm_backend"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"